+++ Sources/MapboxNavigation/Dictionary.swift
@@ -1,16 +1,15 @@
 import Foundation
 
-extension Dictionary where Key == Int, Value: NSExpression {
+extension Dictionary where Key == Double, Value == Double {
     /**
      Returns a copy of the stop dictionary with each value multiplied by the given factor.
      */
     public func multiplied(by factor: Double) -> Dictionary {
-        var newCameraStop: [Int: NSExpression] = [:]
+        var newCameraStop: [Double: Double] = [:]
         for stop in self {
-            let currentValue = stop.value.constantValue as! Double
-            let newValue =  currentValue * factor
-            newCameraStop[stop.key] = NSExpression(forConstantValue: newValue)
+            let newValue =  stop.value * factor
+            newCameraStop[stop.key] = newValue
         }
-        return newCameraStop as! Dictionary<Key, Value>
+        return newCameraStop
     }
 }