+++ Sources/MapboxNavigation/Expression.swift
@@ -1,0 +1,14 @@
+import MapboxMaps
+
+extension Expression {
+    
+    static func routeLineWidthExpression(_ multiplier: Double = 1.0) -> Expression {
+        let lineWidthExpression = Exp(.interpolate) {
+            Exp(.linear)
+            Exp(.zoom)
+            RouteLineWidthByZoomLevel.multiplied(by: multiplier)
+        }
+        
+        return lineWidthExpression
+    }
+}